requirements.txt
bs4
jsonlines
jsonschema
openpyxl
pycrypto
pyOpenSSL
requests
robotframework==3.0.4
robotframework-faker
robotframework-requests
robotframework-sshlibrary
urllib3
